Abtract util, add comments
